+++ hack1/papers/metadata.csv
@@ -1,0 +1,6 @@
+title,authors,abstract,year,pdf_url,publisher
+Synthetic accessibility and stability rules of NASICONs,"Bin Ouyang ,Jingyang Wang ,Tanjin He ,Christopher J. Bartel ,Haoyan Huo ,Yan Wang ,Valentina Lacivita ,Haegyeom Kim ,Gerbrand Ceder","In this paper we develop the stability rules for NASICON structured materials, as an example of compounds with complex bond topology and composition. By applying machine learning to the ab-initio computed phase stability of 3881 potential NASICONs we can extract a simple two-dimensional descriptor that is extremely good at separating stable from unstable NASICONS. This machine-learned ""tolerance factor"" contains information on the Na content, the radii and electronegativities of the elements, and the Madelung energy. We test the predictive capability of this approach by selecting six predicted NASICON compositions. Five out of the six resulted in a phase pure NASICON while the sixth composition led to a NASICON that coexisted with other phases, validating the efficacy of this approach. This work not only provide tools to understand synthetic accessibility of NASICON-type materials, but also demonstrate an efficient paradigm for discovering new materials with complicate composition and atomic structure.",2021.0,http://arxiv.org/pdf/2102.03627v2,arXiv
+Deep learning-driven evaluation and prediction of ion-doped NASICON materials for enhanced solid-state battery performance,"Zirui Zhao ,Xiaoke Wang ,Si Wu ,Pengfei Zhou ,Qian Zhao ,Guanping Xu ,Kaitong Sun ,Hai-Feng Li","We developed a convolutional neural network (CNN) model capable of predicting the performance of various ion-doped NASICON compounds by leveraging extensive datasets from prior experimental investigation.The model demonstrated high accuracy and efficiency in predicting ionic conductivity and electrochemical properties. Key findings include the successful synthesis and validation of three NASICON materials predicted by the model, with experimental results closely matching the model predictions. This research not only enhances the understanding of ion-doping effects in NASICON materials but also establishes a robust framework for material design and practical applications. It bridges the gap between theoretical predictions and experimental validations.",2024.0,http://arxiv.org/pdf/2409.02952v1,arXiv
+Kinetic Monte Carlo Simulations of Sodium Ion Transport in NaSICON Electrodes,"Ziliang Wang ,Tara P. Mishra ,Weihang Xie ,Zeyu Deng ,Gopalakrishnan Sai Gautam ,Anthony K. Cheetham ,Pieremanuele Canepa","The development of high-performance sodium (Na) ion batteries requires improved electrode materials. The energy and power densities of Na superionic conductor (NaSICON) electrode materials are promising for large-scale energy storage applications. However, several practical issues limit the full utilization of the theoretical energy densities of NaSICON electrodes. A pressing challenge lies in the limited sodium extraction in low Na content NaSICONs, e.g., $\rm Na_1V^{IV}V^{IV}(PO_4)_3 \leftrightarrow V^{V}V^{IV}(PO_4)_3 + 1e^- + 1Na^+$. Hence, it is important to quantify the Na-ion mobility in a broad range of NaSICON electrodes. Using a kinetic Monte Carlo approach bearing the accuracy of first-principles calculations, we elucidate the variability of Na-ion transport vs. Na content in three important NaSICON electrodes, Na$_{\rm x}$Ti$_{2}$(PO$_{4}$)$_{3}$, Na$_{\rm x}$V$_{2}$(PO$_{4}$)$_{3}$, and Na$_{\rm x}$Cr$_{2}$(PO$_{4}$)$_{3}$. Our study suggests that Na$^+$ transport in NaSICON electrodes is almost entirely determined by the local electrostatic and chemical environment set by the transition metal and the polyanionic scaffold. The competition with the ordering-disordering phenomena of Na-vacancies also plays a role in influencing Na-transport. We link the variations in the Na$^+$ kinetic properties by analyzing the competition of ligand field stabilization transition metal ions and their ionic radii. We interpret the limited Na-extraction at $x = 1$ observed experimentally by gaining insights into the local Na-vacancy interplay. We propose that targeted chemical substitutions of transition metals disrupting local charge arrangements will be critical to reducing the occurrence of strong Na$^+$-vacancy orderings at low Na concentrations, thus, expanding the accessible capacities of these electrode materials.",2023.0,http://arxiv.org/pdf/2308.04772v1,arXiv
+MgF$_2$ as an effective additive for improving ionic conductivity of ceramic solid electrolytes,"Pengfei Zhou ,Kaitong Sun ,Shunping Ji ,Zirui Zhao ,Ying Fu ,Junchao Xia ,Si Wu ,Yinghao Zhu ,Kwun Nam Hui ,Hai-Feng Li","As typical solid-state electrolytes (SSEs), {Na}$_{1+x}${Zr}$_2${Si}$_{x}${P}$_{3-x}${O}$_{12}$ NASICONs provide an ideal platform for solid-state batteries (SSBs) that display higher safety and accommodate higher energy densities. The critical points for achieving SSBs with higher efficiencies are to improve essentially the ionic conductivity and to reduce largely the interfacial resistance between SSEs and cathode materials, which would necessitate extremely high level of craftsmanship and high-pressure equipment. An alternative to higher-performance and lower-cost SSBs is additive manufacturing. Here, we report on an effective additive, MgF$_2$, which was used in synthesizing NASICONs, resulting in SSEs with fewer defects and higher performance. With an addition of mere 1 wt$\%$ MgF$_2$ additive, the total room-temperature ionic conductivity of the NASICON electrolyte reaches up to 2.03 mS cm$^{-1}$, improved up to $\sim$ 181.3$\%$, with an activation energy of 0.277 eV. Meanwhile, the stability of the Na plating/stripping behavior in symmetric cells increases from 236 to 654 h. We tried to reveal the microscopic origins of the higher ionic conductivity of MgF$_2$-doped NASICONs by comprehensive in-house characterizations. Our study discovers a novel MgF$_2$ additive and provides an efficient way to prepare higher-performance SSEs, making it possible to fabricate lower-cost SSBs in industries.",2023.0,http://arxiv.org/pdf/2302.07264v1,arXiv
+Phase Stability and Sodium-Vacancy Orderings in a NaSICON Electrode,"Ziliang Wang ,Sunkyu Park ,Zeyu Deng ,Dany Carlier ,Jean-Noël Chotard ,Laurence Croguennec ,Gopalakrishnan Sai Gautam ,Anthony K. Cheetham ,Christian Masquelier ,Pieremanuele Canepa","We elucidate the thermodynamics of sodium (Na) intercalation into the sodium super-ionic conductor (NaSICON)-type electrode, Na$_x$V$_2$(PO$_4$)$_3$, for promising Na-ion batteries with high-power density. This is the first report of a computational temperature-composition phase diagram of the NaSICON-type electrode Na$_x$V$_2$(PO$_4$)$_3$. We identify two thermodynamically stable phases at the compositions Na$_2$V$_2$(PO$_4$)$_3$ and Na$_{3.5}$V$_2$(PO$_4$)$_3$, and their structural features are described for the first time based on our computational analysis. We unveil the crystal-structure and the electronic-structure origins of the ground-state compositions associated with specific Na/vacancy arrangements, which are driven by charge orderings on the vanadium sites. These results are significant for the optimization of high-energy and power densities electrodes for sustainable Na-ion batteries",2021.0,http://arxiv.org/pdf/2109.06997v1,arXiv
